# SYMFONY_API_Voluntariado4v/api_voluntariado_4v/openapi.yaml
openapi: 3.0.0
info:
  title: 'API Voluntariado 4V'
  description: 'API REST para la gesti├│n de voluntariado escolar (Alumnos, Coordinadores y ONGs).'
  version: 1.0.0
paths:
...
  '/organizaciones/{id}/actividades/{actividadId}/voluntarios':
    get:
      tags:
        - Organizaciones
      operationId: get_api_voluntarios_actividad_organizacion
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
        -
          name: actividadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Lista de voluntarios inscritos en la actividad'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { id_voluntario: { type: integer }, nombre: { type: string }, apellidos: { type: string }, email: { type: string }, fecha_solicitud: { type: string, format: date-time }, estado_solicitud: { type: string, enum: [Pendiente, Confirmada, Rechazada] } }
                  type: object
...
  '/usuarios/{id}/rol':
    put:
      tags:
        - Usuarios
      operationId: put_api_cambiar_rol_usuario
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: 'Nuevo rol para el usuario'
        required: true
        content:
          application/json:
            schema:
              properties:
                id_rol:
                  description: '1=Coordinador, 2=Organizaci├│n, 3=Voluntario'
                  type: integer
                  example: 2
              type: object
      responses:
        '200':
          description: 'Rol actualizado correctamente'
          content:
            application/json:
              schema:
                properties:
                  mensaje: { type: string }
                  usuario: { type: object }
                  advertencia: { type: string, nullable: true }
                type: object
        '404':
          description: 'Usuario o Rol no encontrado'
        '400':
          description: 'Datos inv├ílidos'
  /voluntarios:
    get:
      tags:
        - Voluntarios
      operationId: get_api_listar_voluntarios
      responses:
        '200':
          description: 'Listado completo de voluntarios activos'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { id_usuario: { type: integer, example: 10 }, nombre: { type: string, example: Ana }, apellidos: { type: string, example: Garc├¡a }, correo_usuario: { type: string, example: ana@test.com }, curso_actual: { type: string, example: '2┬║ DAW' }, telefono: { type: string, example: '+34 600 11 22 33' }, estado_cuenta: { type: string, example: Activa } }
                  type: object
    post:
      tags:
        - Voluntarios
      operationId: post_api_registro_voluntario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoluntarioCreateDTO'
      responses:
        '201':
          description: 'Voluntario registrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoluntarioResponseDTO'
  '/voluntarios/{id}':
    get:
      tags:
        - Voluntarios
      operationId: get_api_get_voluntario
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Detalle del voluntario'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoluntarioResponseDTO'
    put:
      tags:
        - Voluntarios
      operationId: put_api_actualizar_voluntario
      parameters:
        -
          name: X-User-Id
          in: header
          description: 'ID del usuario logueado'
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoluntarioUpdateDTO'
      responses:
        '200':
          description: 'Perfil actualizado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoluntarioResponseDTO'
  '/voluntarios/{id}/actividades/{idActividad}':
    post:
      tags:
        - Voluntarios
      operationId: post_api_inscribirse_actividad
      parameters:
        -
          name: X-User-Id
          in: header
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: string
        -
          name: idActividad
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 'Inscripci├│n realizada'
        '409':
          description: 'Error de reglas de negocio (cupo, fechas, duplicado)'
    delete:
      tags:
        - Voluntarios
      operationId: delete_api_desapuntarse_actividad
      parameters:
        -
          name: X-User-Id
          in: header
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: string
        -
          name: idActividad
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Te has desapuntado correctamente'
        '404':
          description: 'No estas inscrito en esta actividad'
  '/voluntarios/{id}/historial':
    get:
      tags:
        - Voluntarios
      operationId: get_api_historial_voluntario
      parameters:
        -
          name: X-User-Id
          in: header
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Historial detallado y resumen'
          content:
            application/json:
              schema:
                properties:
                  resumen: { type: object }
                  actividades: { type: array, items: { $ref: '#/components/schemas/InscripcionResponseDTO' } }
                type: object
  '/voluntarios/{id}/recomendaciones':
    get:
      tags:
        - Voluntarios
      operationId: get_api_recomendaciones_voluntario
      parameters:
        -
          name: X-User-Id
          in: header
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Lista de actividades recomendadas'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { id_actividad: { type: integer, example: 12 }, titulo: { type: string, example: 'Actividad Recomendada' }, compatibilidad: { type: integer, example: 100 }, motivo: { type: string, example: 'Coincide con tus ODS y disponibilidad' } }
                  type: object
  '/voluntarios/{id}/horas-totales':
    get:
      tags:
        - Voluntarios
      operationId: get_api_horas_totales_voluntario
      parameters:
        -
          name: X-User-Id
          in: header
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Total de horas de voluntariado realizadas (campo calculado)'
          content:
            application/json:
              schema:
                properties:
                  horas_totales: { description: 'Suma total de horas de actividades Aceptadas o Finalizadas', type: integer, example: 45 }
                type: object
...
components:
  schemas:
...
    VoluntarioCreateDTO:
      required:
        - google_id
        - correo
        - nombre
        - apellidos
        - dni
        - telefono
        - fecha_nac
        - carnet_conducir
        - id_curso_actual
      properties:
        google_id:
          type: string
          example: google_usr_99
        correo:
          type: string
          example: voluntario@test.com
        nombre:
          type: string
          minLength: 2
          example: Ana
        apellidos:
          type: string
          example: Garc├¡a
        dni:
          type: string
          example: 12345678Z
        telefono:
          type: string
          example: '+34 655 44 33 22'
        fecha_nac:
          type: string
          example: '1999-12-31'
        carnet_conducir:
          type: boolean
          example: true
        id_curso_actual:
          type: integer
          example: 2
        descripcion:
          type: string
          default: null
          maxLength: 500
          example: 'Me gusta ayudar con temas de log├¡stica.'
          nullable: true
        preferencias_ids:
          type: array
          items:
            type: integer
          default: []
          example:
            - 1
            - 3
        idiomas:
          type: array
          items:
            properties:
              id_idioma:
                type: integer
              nivel:
                type: string
          default: []
          example:
            -
              id_idioma: 1
              nivel: B2
      type: object
...
    VoluntarioResponseDTO:
      required:
        - id_usuario
        - nombre
        - apellidos
        - nombre_completo
        - correo
        - carnet_conducir
        - curso
        - estado_cuenta
        - preferencias
        - idiomas
      properties:
        id_usuario:
          type: integer
          example: 10
        nombre:
          type: string
          example: Ana
        apellidos:
          type: string
          example: Garc├¡a
        nombre_completo:
          type: string
          example: 'Ana Garc├¡a'
        correo:
          type: string
          example: ana@test.com
        img_perfil:
          type: string
          example: usr_10_pic.jpg
          nullable: true
        dni:
          type: string
          example: 12345678Z
          nullable: true
        telefono:
          type: string
          example: '+34 600 00 00 00'
          nullable: true
        fecha_nac:
          type: string
          example: '1995-05-20'
          nullable: true
        carnet_conducir:
          type: boolean
          example: true
        id_curso:
          type: integer
          example: 2
          nullable: true
        curso:
          type: string
          example: '2┬║ DAM'
        estado_cuenta:
          type: string
          example: Activa
        descripcion:
          type: string
          example: 'Me encanta participar en actvidades sociales.'
          nullable: true
        preferencias:
          description: 'Tipos de voluntariado preferidos'
          type: array
          items:
            type: string
            example: Social
        idiomas:
          description: 'Idiomas del voluntario'
          type: array
          items:
            properties:
              id_idioma:
                type: integer
                example: 1
              idioma:
                type: string
                example: Ingl├®s
              nivel:
                type: string
                example: B2
      type: object
  securitySchemes:
    Bearer:
      type: http
      bearerFormat: JWT
      scheme: bearer
...